and then, pt_BR.lang was born.
DirtChest 9000 is a copyrighted product name, so it's better to leave it as is :P
I didn't found a fitting word for upgrade, so I left it that way. It's very likely that whoever see it will understand, though. Not an alien word, IYKWIM.
As for the upgrades, "Upgrade de Baú de Diamante para Baú de Obsidiana" simply doesn't fit in my mind, so I did the "Chest Upgrade:" thingy. If it's a bother, you have the final word.